+++ src/storage/invertedindex/column_inverter.cppm
@@ -6,13 +6,43 @@ import column_vector;
 import parser;
 
 namespace infinity {
+
+class RefCount {
+    std::mutex lock_;
+    std::condition_variable cv_;
+    u32 ref_count_;
+
+public:
+    RefCount();
+    virtual ~RefCount();
+    void Retain() noexcept;
+    void Release() noexcept;
+    void WaitForZeroRefCount();
+    bool ZeroRefCount();
+};
+
+export struct InverterReference {
+
+    void Retain() { ref_count_.Retain(); }
+
+    void Release() { ref_count_.Release(); }
+
+    bool ZeroRefCount() { return ref_count_.ZeroRefCount(); }
+
+    void WaitForZeroRefCount() { return ref_count_.WaitForZeroRefCount(); }
+
+    RefCount &GetRefCount() { return ref_count_; }
+
+    virtual void Commit() = 0;
+
+    RefCount ref_count_;
+};
+
 export class ColumnInverter {
 public:
     virtual void InvertColumn(SharedPtr<ColumnVector> column_vector, Vector<RowID> &row_ids) = 0;
 
     virtual void InvertColumn(u32 doc_id, const String &val) = 0;
-
-    virtual void Commit() = 0;
 
     virtual void Flush() = 0;
 };
