+++ src/storage/invertedindex/memory_indexer.cppm
@@ -111,9 +111,9 @@ private:
     Deque<UniquePtr<SequentialColumnInverter>> inflight_inverters_;
     UniquePtr<SequentialColumnInverter> inverter_;
 
-    Vector<UniquePtr<ParallelColumnInverter>> parallel_inverters_;
-    Vector<Vector<UniquePtr<ParallelColumnInverter>>> free_parallel_inverters_;
-    Deque<Vector<UniquePtr<ParallelColumnInverter>>> inflight_parallel_inverters_;
+    UniquePtr<ParallelColumnInverters> parallel_inverter_;
+    Vector<UniquePtr<ParallelColumnInverters>> free_parallel_inverters_;
+    Deque<UniquePtr<ParallelColumnInverters>> inflight_parallel_inverters_;
 
     u32 num_inverters_;
     u32 max_inverters_;
